Rétablissement de la carte ampleur de prod
On reviendra à AmpleurSummary depuis SimplifiedAmpleurSummary dans une
prochaine PR
