--- components/Result.tsx
@@ -1,250 +1,0 @@
-import { Labels } from '@/app/LandingUI'
-import css from '@/components/css/convertToJs'
-import { formatValue } from '@/node_modules/publicodes/dist/index'
-import Link from 'next/link'
-import { PrimeStyle } from './Geste'
-import { CTA, CTAWrapper, Card, cardBorder } from './UI'
-import { ExplicationMPRA, ExplicationMPRG } from './explications/Éligibilité'
-import GestesPreview from './mprg/GestesPreview'
-import AidesLocales from './AidesLocales'
-
-/* This component was first written for simulation mode where the state could be success, running or fail. Since then we've switched to a more classic result where it
- * can only be success or fail. I've kept this object for future references, for its colors */
-export const colors = {
-  success: {
-    background: '#297254',
-    lightBackground: '#c4fad5', //TODO use this for the contour, more beautiful
-    label: 'Estimation finale',
-  },
-  running: {
-    background: 'var(--color)',
-    color: 'white',
-    label: 'Sous conditions',
-  },
-  fail: {
-    background: 'salmon',
-    lightBackground: '#f6b7af',
-    color: 'white',
-    label: 'Non éligible',
-  },
-  //  waiting: { background: '#9f9f9f', color: 'white', label: 'À suivre' },
-}
-
-export default function Result({
-  engine,
-  isFinal,
-  rules,
-  dottedName,
-  hideNumeric = false,
-  index,
-  url,
-  situation,
-}) {
-  const rule = rules[dottedName]
-  const evaluation = engine
-    .setSituation({ ...situation, 'projet . travaux': 999999 })
-    .evaluate(dottedName)
-  console.log('result', evaluation)
-
-  const value = formatValue(evaluation, { precision: 0 })
-  const isNotApplicable =
-    value === 'Non applicable' || evaluation.nodeValue === 0 || value === 'non'
-
-  const state = isNotApplicable ? 'fail' : isFinal ? 'success' : 'running',
-    fail = state === 'fail'
-  const { color, background, label } = colors[state]
-  const MPRA = dottedName === 'MPR . accompagnée . montant'
-
-  return (
-    <li
-      css={`
-        margin: 0;
-        width: 22rem;
-        max-width: min(22rem, 90%);
-        @media (min-width: 800px) {
-          width: 100%;
-          max-width: initial;
-        }
-        border-color: ${fail ? '#ddd' : '#dfdff1'};
-
-        position: relative;
-
-        h3 {
-          font-size: 120%;
-          margin: 2rem 0 0.4rem;
-        }
-        @media (min-width: 800px) {
-        }
-      `}
-    >
-      {' '}
-      {MPRA ? (
-        <section>
-          <header>
-            <h3>Faire un rénovation d'ampleur</h3>
-            <Labels
-              $color={'#6E4444'}
-              $background={'#fdf8db'}
-              css={`
-                margin-top: 0.3rem;
-              `}
-            >
-              {['🤝 Un professionnel vous accompagne'].map((text) => (
-                <li key={text}>{text}</li>
-              ))}
-            </Labels>
-            <p
-              css={`
-                margin-top: 1.4rem;
-              `}
-            >
-              Un programme sur-mesure pour gagner au minimum{' '}
-              deux&nbsp;classes&nbsp;DPE.
-            </p>
-          </header>
-          <Card
-            css={`
-              color: ${fail ? '#888' : 'inherit'};
-              margin-top: 0.2rem;
-              background: white;
-            `}
-          >
-            <h4
-              style={css`
-                font-weight: 400;
-                margin: 1rem 0 0rem;
-                font-size: 120%;
-              `}
-              dangerouslySetInnerHTML={{ __html: rule.titreHtml }}
-            />
-
-            {fail ? (
-              <div
-                css={`
-                  margin: 1rem 0;
-                  color: black;
-                  text-align: center;
-                `}
-              >
-                <ExplicationMPRA {...{ engine, situation }} />
-              </div>
-            ) : (
-              <PrimeStyle $inactive={fail}>
-                {isFinal ? `` : fail ? `` : `Jusqu'à `} <strong>{value}</strong>
-              </PrimeStyle>
-            )}
-
-            {/* On suppose pour l'instant que toutes les aides locales sont pour des rénovations d'ampleur, mais ce ne sera pas le cas ! */}
-            <AidesLocales {...{ engine, situation }} />
-            <div
-              css={`
-                visibility: ${!isNotApplicable && url ? 'visible' : 'hidden'};
-                > div {
-                  margin-bottom: 0.3rem;
-                  margin-top: 1rem;
-                }
-              `}
-            >
-              <CTAWrapper $justify="start">
-                <CTA $fontSize="normal">
-                  <Link href={url}>
-                    {MPRA
-                      ? 'Découvrir le détail'
-                      : 'Voir les 20 gestes disponibles'}
-                  </Link>
-                </CTA>
-              </CTAWrapper>
-            </div>
-          </Card>
-        </section>
-      ) : (
-        <section>
-          <header>
-            <h3>Rénover à la carte</h3>
-            <p
-              css={`
-                margin-top: 1.4rem;
-              `}
-            >
-              Rénovez progressivement via un bouquet de gestes subventionnés.
-            </p>
-          </header>
-          <Card
-            css={`
-              color: ${fail ? '#888' : 'inherit'};
-            `}
-          >
-            <h4
-              style={css`
-                font-weight: 400;
-                margin: 1rem 0 0rem;
-                font-size: 120%;
-              `}
-              dangerouslySetInnerHTML={{ __html: rule.titreHtml }}
-            />
-            {fail && (
-              <div
-                css={`
-                  margin: 1rem 0;
-                  color: black;
-                  text-align: center;
-                `}
-              >
-                <ExplicationMPRG {...{ engine, situation }} />
-              </div>
-            )}
-
-            {!fail ? (
-              <GestesPreview
-                {...{
-                  rules,
-                  inactive: fail,
-                  dottedNames: [
-                    'gestes . recommandés . audit',
-                    'gestes . chauffage . PAC . air-eau',
-                    'gestes . isolation . murs extérieurs',
-                  ],
-                  engine,
-                  situation,
-                }}
-              />
-            ) : (
-              <span>
-                <a
-                  target="_blank"
-                  href="https://www.service-public.fr/particuliers/vosdroits/F35083"
-                >
-                  En savoir plus sur ce parcours
-                </a>
-                .
-              </span>
-            )}
-            <div
-              css={`
-                display: ${!isNotApplicable && url ? 'visible' : 'none'};
-                > div {
-                  margin-bottom: 0.3rem;
-                  margin-top: 1.6rem;
-                }
-              `}
-            >
-              <CTAWrapper $justify="start">
-                <CTA $fontSize="normal">
-                  <Link href={url}>
-                    {MPRA ? (
-                      'Découvrir le détail'
-                    ) : (
-                      <span>
-                        Voir les <strong>20</strong> gestes disponibles
-                      </span>
-                    )}
-                  </Link>
-                </CTA>
-              </CTAWrapper>
-            </div>
-          </Card>
-        </section>
-      )}
-    </li>
-  )
-}
